+++ app/src/main/java/com/vinylsmobile/data/api/AlbumApi.kt
@@ -1,6 +1,6 @@
 package com.vinylsmobile.data.api
 
-import com.vinylsmobile.data.model.Album
+import com.vinylsmobile.model.Album
 import retrofit2.http.GET
 
 interface AlbumApi {
