+++ app/src/main/java/com/vinylsmobile/model/Album.kt
@@ -1,4 +1,4 @@
-package com.vinylsmobile.data.model
+package com.vinylsmobile.model
 
 data class Album (
     val id: Int,
