+++ app/src/main/java/com/vinylsmobile/model/Performer.kt
@@ -1,4 +1,4 @@
-package com.vinylsmobile.data.model
+package com.vinylsmobile.model
 
 data class Performer(
     val id: Int,
