+++ app/src/main/java/com/vinylsmobile/view/AlbumFragment.kt
@@ -1,23 +1,17 @@
-package com.vinylsmobile.ui.fragments
+package com.vinylsmobile.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.vinylsmobile.data.api.AlbumApi
-import com.vinylsmobile.data.repository.AlbumRepository
+import com.vinylsmobile.repository.AlbumRepository
 import com.vinylsmobile.databinding.FragmentAlbumBinding
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import com.vinylsmobile.ui.adapters.AlbumAdapter
-import com.vinylsmobile.ui.viewmodels.AlbumViewModel
-import com.vinylsmobile.ui.viewmodels.AlbumViewModelFactory
-import kotlinx.coroutines.launch
+import com.vinylsmobile.view.adapters.AlbumAdapter
+import com.vinylsmobile.viewmodels.AlbumViewModel
+import com.vinylsmobile.viewmodels.AlbumViewModelFactory
 
 class AlbumFragment : Fragment() {
     private var _binding: FragmentAlbumBinding? = null
