+++ app/src/main/res/layout/activity_collection.xml
@@ -5,7 +5,7 @@
     android:id="@+id/collection"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.vinylsmobile.activities.MainActivity">
+    tools:context="com.vinylsmobile.view.MainActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
